Various refactorings and bug-fixes.
1. Changed how padding is implemented so that padding is only added to
the coefficient array before the wavelet transform, and is removed
from the coefficient array immediately after the wavelet transform.
No metadata for "OrigXl" etc is needed now.

2. Various places "frame" is changed to "picture" and "fparams" to
"pparams" etc.

3. Fixed bug with misreporting bit rate for long sequences by recording
data with int64_t not int.

4. Fixed a bug with perceptual weighting where additional weight was being
given to the DC band even though it is coded losslessly. This threw out
weights for the other bands when the weights were normalised, especially
for intra pictures. In turn this required that the relative lambdas used
for different picture types had to be changed.

5. Fixed a minor bug with rate control where the intra lambda was not being
applied correctly.

6. Changed "field coding" metadata to "picture coding mode" metadata
throughout, as per spec.

7. Changed "QIndex" throughout to "QuantIndex" etc for clarity.

8. Refactored picture coding and decoding to simplify things a little.

+++ libdirac_byteio/dirac_byte_stats.cpp
@@ -60,14 +60,14 @@ void DiracByteStats::Clear()
     m_byte_count.clear();
 }
 
-int DiracByteStats::GetBitCount(const StatType& stat_type) const  
+int64_t DiracByteStats::GetBitCount(const StatType& stat_type) const  
 {
     return GetByteCount(stat_type) * CHAR_BIT;
 }
 
-int DiracByteStats::GetByteCount(const StatType& stat_type) const
+int64_t DiracByteStats::GetByteCount(const StatType& stat_type) const
 {
-     std::map<StatType, int>::const_iterator it;
+     std::map<StatType, long int>::const_iterator it;
      it = m_byte_count.find(stat_type);
      if(it==m_byte_count.end())
            return 0;
@@ -75,7 +75,7 @@ int DiracByteStats::GetByteCount(const StatType& stat_type) const
       return it->second;
 }
 
-void DiracByteStats::SetByteCount(const StatType& stat_type, int count)
+void DiracByteStats::SetByteCount(const StatType& stat_type, int64_t count)
 {
     if(m_byte_count.find(stat_type)==m_byte_count.end())
         m_byte_count[stat_type]=0;
